requirements.txt
streamlit==1.38.0
sentence-transformers==3.0.1
transformers==4.44.0
faiss-cpu==1.8.0
torch==2.3.1
numpy==1.26.4
